fix: use combine reducers, simplify useTextEditorState
