+++ packages/picasso-lab/src/TextEditor/types.ts
@@ -1,8 +1,5 @@
 import { MouseEventHandler, ChangeEvent } from 'react'
 
-import { EditorActionsType, EditorStateType } from './store/editor/types'
-import { ToolbarActionsType, ToolbarStateType } from './store/toolbar/types'
-
 export type HTMLString = string
 
 export type HeaderValueType = '3' | ''
@@ -16,10 +13,3 @@ export type SelectOnChangeHandler = (
     value: HeaderValueType
   }>
 ) => void
-
-export type StateType = {
-  toolbar: ToolbarStateType
-  editor: EditorStateType
-}
-
-export type ActionsType = ToolbarActionsType | EditorActionsType
